# jerrinot's improvement - fast-path for short keys