#### Check List:
- [x] Tests pass (`./test.sh <username>` shows no differences between expected and actual outputs)
- [x] All formatting changes by the build are committed
- [x] Your launch script is named `calculate_average_<username>.sh` (make sure to match casing of your GH user name) and is executable
- [x] Output matches that of `calculate_average_baseline.sh`
- [ ] For new entries, or after substantial changes: When implementing custom hash structures, please point to where you deal with hash collisions (line number)

* Execution time: 2.6s
* Execution time of reference implementation: looong

before:
```
Performance counter stats for './calculate_average_jerrinot.sh':

         18,744.21 msec task-clock                       #    5.816 CPUs utilized             
             5,311      context-switches                 #  283.341 /sec                      
                97      cpu-migrations                   #    5.175 /sec                      
           230,736      page-faults                      #   12.310 K/sec                     
    53,310,529,541      cycles                           #    2.844 GHz                         (40.83%)
       146,723,565      stalled-cycles-frontend          #    0.28% frontend cycles idle        (40.74%)
     1,698,491,317      stalled-cycles-backend           #    3.19% backend cycles idle         (40.57%)
    99,057,621,741      instructions                     #    1.86  insn per cycle            
                                                  #    0.02  stalled cycles per insn     (40.24%)
    11,152,157,044      branches                         #  594.965 M/sec                       (39.98%)
       607,305,255      branch-misses                    #    5.45% of all branches             (39.82%)
    35,747,805,369      L1-dcache-loads                  #    1.907 G/sec                       (39.76%)
     1,395,129,518      L1-dcache-load-misses            #    3.90% of all L1-dcache accesses   (39.90%)
   <not supported>      LLC-loads                                                             
   <not supported>      LLC-load-misses                                                       
     2,374,784,638      L1-icache-loads                  #  126.694 M/sec                       (40.02%)
         1,965,683      L1-icache-load-misses            #    0.08% of all L1-icache accesses   (40.13%)
       652,714,679      dTLB-loads                       #   34.822 M/sec                       (40.30%)
         5,492,098      dTLB-load-misses                 #    0.84% of all dTLB cache accesses  (40.48%)
         3,381,604      iTLB-loads                       #  180.408 K/sec                       (40.63%)
           377,331      iTLB-load-misses                 #   11.16% of all iTLB cache accesses  (40.69%)
       516,907,135      L1-dcache-prefetches             #   27.577 M/sec                       (40.68%)
   <not supported>      L1-dcache-prefetch-misses                                             

       3.222922313 seconds time elapsed

      16.403446000 seconds user
       2.143290000 seconds sys
```

after:
```
 Performance counter stats for './calculate_average_jerrinot.sh':

         17,056.68 msec task-clock                       #    5.778 CPUs utilized             
             5,475      context-switches                 #  320.989 /sec                      
               197      cpu-migrations                   #   11.550 /sec                      
           232,801      page-faults                      #   13.649 K/sec                     
    48,121,960,155      cycles                           #    2.821 GHz                         (40.54%)
       173,699,958      stalled-cycles-frontend          #    0.36% frontend cycles idle        (40.63%)
     2,058,659,332      stalled-cycles-backend           #    4.28% backend cycles idle         (40.75%)
   108,222,728,495      instructions                     #    2.25  insn per cycle            
                                                  #    0.02  stalled cycles per insn     (40.58%)
     9,480,191,513      branches                         #  555.805 M/sec                       (40.45%)
       194,057,814      branch-misses                    #    2.05% of all branches             (40.23%)
    34,692,557,868      L1-dcache-loads                  #    2.034 G/sec                       (40.51%)
     1,066,166,776      L1-dcache-load-misses            #    3.07% of all L1-dcache accesses   (40.48%)
   <not supported>      LLC-loads                                                             
   <not supported>      LLC-load-misses                                                       
     2,252,492,384      L1-icache-loads                  #  132.059 M/sec                       (40.32%)
         2,298,014      L1-icache-load-misses            #    0.10% of all L1-icache accesses   (40.44%)
       686,391,714      dTLB-loads                       #   40.242 M/sec                       (40.32%)
         5,115,893      dTLB-load-misses                 #    0.75% of all dTLB cache accesses  (40.17%)
         4,159,252      iTLB-loads                       #  243.849 K/sec                       (39.88%)
           483,889      iTLB-load-misses                 #   11.63% of all iTLB cache accesses  (39.91%)
       321,355,018      L1-dcache-prefetches             #   18.840 M/sec                       (40.42%)
   <not supported>      L1-dcache-prefetch-misses                                             

       2.952213866 seconds time elapsed

      14.534809000 seconds user
       2.319250000 seconds sys
```

The main difference: segregated map for short keys.