@gunnarmorling 2024-01-21
Does this match the expected output for a 1B file for you?