@jerrinot 2024-01-21
should be fixed now. the fix cost some ms, but it still should be quite a bit faster than before.