@gunnarmorling 2024-01-23
00:02.374 now, nice!